Starting state of CSharp Excerise one
