+++ CSharp/MyApplication/Program.cs
@@ -1,62 +1,50 @@
-﻿namespace School {
-    class GradesSet{
-        public GradesSet(){}
-        public GradesSet(double _E, double _M){
-            _english = _E;
-            _maths = _M;
-            _avgGrade = calculateAvgGrade();
-        }
+using System;
+using System.Collections.Generic;
 
-        double _english;
-        double _maths;
-        double _avgGrade;
-
-        public double English { get => _english; set => _english = value; }
-        public double Maths { get => _maths; set => _maths = value; }
-        public double AvgGrade { get => _avgGrade; set => _avgGrade = value; }
-
-        private double calculateAvgGrade() { return (_english + _maths) * 0.5; }
-        
-    }
-
-    class Student {
-        public Student(string _F, string _L, int _A, GradesSet _G) {
-            _firstName = _F;
-            _lastName = _L;
-            _age = _A;
-            _grades = _G;
-        }
-
-        string _firstName = "default";
-        string _lastName = "default";
-        int _age = 0;
-
-        GradesSet _grades;
-
-        public string FirstName { get => _firstName; set => _firstName = value; }
-        public string LastName { get => _lastName; set => _lastName = value; }
-        public int Age { get => _age; set => _age = value; }
-        internal GradesSet Grades { get => _grades; set => _grades = value; }
-
-        public void printStudent() { Console.WriteLine(_firstName + "\n" + _lastName + "\n" + _age + "\n" + Grades.AvgGrade); }
-
-    }
-}
-
-
-namespace MyApplication
+class Program
 {
-    class Program
+    static void Main(string[] args)
     {
-        static void Main(string[] args)
+        Console.WriteLine("Student Grade Management System");
+        Console.WriteLine("===============================\n");
+
+        // Sample student data
+        List<string> studentNames = new List<string> { "Alex", "Taylor", "Jordan" };
+        List<List<int>> studentScores = new List<List<int>> {
+            new List<int> {95, 88, 92, 85},
+            new List<int> {78, 81, 74, 70},
+            new List<int> {62, 67, 69, 64}
+        };
+
+        // TODO: Calculate averages and print each student's summary
+        for (int i = 0; i < studentScores.Count; i++)
         {
+            string studentName = studentNames[i];
+            List<int> scores = studentScores[i];
 
-            School.Student member = new School.Student("Oliver", "Callaghan", 23, new School.GradesSet(85.4, 89.3));
+            double average = CalculateAverage(scores); // TODO
+            char letterGrade = GetLetterGrade(average); // TODO
 
-            member.Grades.English = 0.9;
-
-            member.printStudent();
-
+            // TODO: Print student details clearly (name, scores, average, letter grade)
         }
+
+        // TODO: Compute and print class statistics:
+        // - Class average
+        // - Highest scoring student
+        // - Lowest scoring student
     }
+
+    // TODO: Method to calculate the average score from a list of integers
+    static double CalculateAverage(List<int> scores)
+    {
+        // Implement calculation here
+    }
+
+    // TODO: Method to determine the letter grade based on average score
+    static char GetLetterGrade(double average)
+    {
+        // Implement logic here
+    }
+
+    // Optional additional methods for class average, highest, and lowest scores
 }
